Added final exam schedule
This commit contains the final exam compressor. This does not compress
the class times as far as would be ideal. this however is what we were
given. Due to only the class start time being the determining factor in
differentiating day classes and night classes, this is pretty much all I
can do with this method of compression.