Added class compression
This commit contains the class compressor. This does not compress class
times as far as would be ideal. This is what we were given to work with.
Due to only the class start time being the deciding factor in
differentiating day classes from night classes. This is pretty much all
I can do with this method of compression.