Compression
Tried plan B and got the Compression down to 21 final classes. I need to
know exactly which classes are allowed to be ignored in plan B as right
now I am ignoring ALL single day classes. I have displayed the warnings
for each of these classes at the top of the console. This gives us
manageable final times, but seems like it might be at too great a cost.